requirements.txt
google-adk
hikari
python-dotenv
platformdirs
pydantic
pytest
pytest-asyncio